# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sales API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obter todos os recursos",
                "operationId": "b63ec4394c475fe1a0fdc836e1e1ed99",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Quantidade de dados por página.",
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "identify",
                        "in": "query",
                        "description": "Filtrar recurso pelo número identificador",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_id",
                        "in": "query",
                        "description": "Filtrar recurso pelo id do usuário",
                        "schema": {
                            "type": "string",
                            "format": "char32"
                        }
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "Filtrar recurso pelo id do cliente",
                        "schema": {
                            "type": "string",
                            "format": "char32"
                        }
                    },
                    {
                        "name": "payment_id",
                        "in": "query",
                        "description": "Filtrar recurso pelo id da forma de pagamento",
                        "schema": {
                            "type": "string",
                            "format": "char32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pedidos retornada com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Armazenar novo recurso no banco de dados",
                "operationId": "b9fc36483260b69b444017c291d48cce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_id": {
                                        "description": "ID do cliente",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
                                    },
                                    "payment_id": {
                                        "description": "ID da forma de pagamento",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "description": "ID do produto",
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
                                                },
                                                "quantity": {
                                                    "description": "Quantidade do produto",
                                                    "type": "integer",
                                                    "minimum": 1,
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Ver um recurso",
                "operationId": "c82bf974ea99be4d0ad4491f2e9a3096",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "char32"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do pedido retornados com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "404": {
                        "description": "Pedido não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Atualizar um recurso no banco de dados",
                "operationId": "d7cc57fbabd8697bc747523009cd61c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "char32"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Pedido atualizado com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "404": {
                        "description": "Pedido não encontrado."
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Deletar recurso",
                "operationId": "ee5586276ea88d598ecfc084ec534427",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "char32"
                        },
                        "examples": {
                            "uuid": {
                                "summary": "UUID value.",
                                "value": "0006faf6-7a61-426c-9034-579f2cfcfa83"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pedido removido com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    },
                    "404": {
                        "description": "Pedido não encontrado."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Listar todos os recursos",
                "description": "Retorna uma lista de recursos disponíveis no banco de dados.",
                "operationId": "getAllResources",
                "responses": {
                    "200": {
                        "description": "Lista de recursos com permissões retornadas com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado, usuário não autenticado."
                    },
                    "403": {
                        "description": "Acesso negado, usuário sem permissão."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}